feat: Implement core news management system with authentication, role-based access, and data persistence.

+++ services/IAccountService.cs
@@ -4,11 +4,11 @@ namespace services;
 
 public interface IAccountService
 {
-    Task<Account?> GetByIdAsync(int id);
+    Task<Account?> GetByIdAsync(short id);
     Task<Account?> GetByEmailAsync(string email);
     Task<List<Account>> GetAllAsync(string? search);
     Task AddAsync(Account account);
     Task UpdateAsync(Account account);
-    Task DeleteAsync(int id);
+    Task DeleteAsync(short id);
     Task<Account?> AuthenticateAsync(string email, string password);
 }
