+++ services/ICategoryService.cs
@@ -4,9 +4,9 @@ namespace services;
 
 public interface ICategoryService
 {
-    Task<Category?> GetByIdAsync(int id);
+    Task<Category?> GetByIdAsync(short id);
     Task<List<Category>> GetAllAsync(string? search);
     Task AddAsync(Category category);
     Task UpdateAsync(Category category);
-    Task DeleteAsync(int id);
+    Task DeleteAsync(short id);
 }
